refactor(core): check alredy built config override
