+++ eodag/api/core.py
@@ -66,7 +66,6 @@ from eodag.utils import (
     MockResponse,
     _deprecated,
     copy_deepcopy,
-    deepcopy,
     get_geometry_from_various,
     makedirs,
     obj_md5sum,
@@ -357,39 +356,6 @@ class EODataAccessGateway:
                 )
                 self.providers_config[provider] = self._pruned_providers_config.pop(
                     provider
-                )
-
-        # check if metada-mapping as already been built as jsonpath in providers_config
-        for provider, provider_conf in conf_update.items():
-            if (
-                provider in self.providers_config
-                and "metadata_mapping" in provider_conf.get("search", {})
-            ):
-                search_plugin_key = "search"
-            elif (
-                provider in self.providers_config
-                and "metadata_mapping" in provider_conf.get("api", {})
-            ):
-                search_plugin_key = "api"
-            else:
-                continue
-            # get some already configured value
-            configured_metadata_mapping = getattr(
-                self.providers_config[provider], search_plugin_key
-            ).metadata_mapping
-            some_configured_value = next(iter(configured_metadata_mapping.values()))
-            # check if the configured value has already been built as jsonpath
-            if (
-                isinstance(some_configured_value, list)
-                and isinstance(some_configured_value[1], tuple)
-                or isinstance(some_configured_value, tuple)
-            ):
-                # also build as jsonpath the incoming conf
-                mtd_cfg_as_conversion_and_querypath(
-                    deepcopy(
-                        conf_update[provider][search_plugin_key]["metadata_mapping"]
-                    ),
-                    conf_update[provider][search_plugin_key]["metadata_mapping"],
                 )
 
         override_config_from_mapping(self.providers_config, conf_update)
